// src/types/entityTypes.ts
export type AssetType = "images" | "map_images";

export type AvailableEntityType =
  | "projects"
  | "manuscripts"
  | "characters"
  | "documents"
  | "maps"
  | "graphs"
  | "dictionaries"
  | "blueprints"
  | "calendars"
  | "timelines"
  | "random_tables"
  | "character_fields_templates"
  | "character_relationship_types"
  | "tags"
  | "conversations"
  | "images";

export type PublicEntities =
  | "characters"
  | "blueprint_instances"
  | "documents"
  | "maps"
  | "graphs"
  | "calendars"
  | "dictionaries"
  | "words";
export type AvailableSubEntityType =
  | "alter_names"
  | "map_pins"
  | "map_layers"
  | "nodes"
  | "edges"
  | "words"
  | "months"
  | "events"
  | "random_table_options"
  | "character_fields"
  | "blueprint_instances"
  | "messages";